Phoenix-rise/requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
supabase>=1.0.3
google-generativeai>=0.3.0
plotly>=5.17.0
pandas>=2.1.0
